update pytest helpers
