+++ backend/python/app/tests/helpers/story_helpers.py
@@ -45,50 +45,3 @@ def guarantee_story_table_not_empty(db):
         )
         db.session.add(obj)
         assert db.session.commit() == None
-
-
-def create_story_translation(db):
-    story = db_session_add_commit_obj(
-        db,
-        Story(
-            title="East of Eden",
-            description="Follow the intertwined destinies of two families whose generations reenact the poisonous rivalry of Cain and Abel.",
-            youtube_link="https://www.youtube.com/watch?v=redECmF7wh8",
-            level=4,
-        ),
-    )
-    translator = db_session_add_commit_obj(
-        db,
-        UserAll(
-            first_name="Carl",
-            last_name="Sagan",
-            email="carl.sagan@uwblueprint.org",
-            auth_id="secret",
-            role="User",
-            approved_languages_translation={"ENGLISH_US": 4, "ENGLISH_UK": 4},
-            approved_languages_review={"ENGLISH_US": 4, "ENGLISH_UK": 4},
-        ),
-    )
-    reviewer = db_session_add_commit_obj(
-        db,
-        UserAll(
-            first_name="Dwight",
-            last_name="D. Eisenhower",
-            email="dwight.d.eisenhower@uwblueprint.org",
-            auth_id="anothersecret",
-            role="User",
-            approved_languages_translation={"ENGLISH_US": 4, "ENGLISH_UK": 4},
-            approved_languages_review={"ENGLISH_US": 4, "ENGLISH_UK": 4},
-        ),
-    )
-    story_translation = db_session_add_commit_obj(
-        db,
-        StoryTranslationAll(
-            story_id=story.id,
-            language="ENGLISH_US",
-            stage="TRANSLATE",
-            translator_id=translator.id,
-            reviewer_id=reviewer.id,
-        ),
-    )
-    return story_translation
