+++ backend/python/app/tests/services/test_story_service.py
@@ -11,7 +11,6 @@ from ..helpers.story_translation_helpers import (
     StoryTranslationRequestDTO,
     assert_story_translation_equals_model,
     create_reviewer,
-    create_story,
     create_story_translation,
     create_story_translation_contents,
     create_translator,
@@ -122,12 +121,12 @@ def test_create_translation_raises_error_if_translation_content_commit_fails_and
 
 
 def test_get_story_translation(app, db, services):
-    translator_obj = create_translator(db)
-    reviewer_obj = create_reviewer(db)
-    story_obj = create_story(db)
-    story_translation_obj = create_story_translation(
-        db, story_obj, translator_obj, reviewer_obj
-    )
+    (
+        _,
+        _,
+        story_obj,
+        story_translation_obj,
+    ) = create_story_translation(db)
     story_translation_contents = create_story_translation_contents(
         db, story_translation_obj
     )
@@ -207,7 +206,7 @@ def test_get_story_translations_available_for_review():
 
 
 def test_remove_reviewer_from_story_translation(app, db, services):
-    story_translation = create_story_translation(db)
+    _, _, _, story_translation = create_story_translation(db)
     assert StoryTranslation.query.get(story_translation.id).reviewer_id != None
     services["story"].remove_user_from_story_translation(
         story_translation_id=story_translation.id, user_id=story_translation.reviewer_id
@@ -216,7 +215,7 @@ def test_remove_reviewer_from_story_translation(app, db, services):
 
 
 def test_remove_translator_from_story_translation(app, db, services):
-    story_translation = create_story_translation(db)
+    _, _, _, story_translation = create_story_translation(db)
     assert StoryTranslation.query.get(story_translation.id).translator_id != None
     services["story"].remove_user_from_story_translation(
         story_translation_id=story_translation.id,
@@ -237,7 +236,7 @@ def test_remove_user_from_invalid_story_translation(app, db, services):
 
 
 def test_remove_invalid_user_from_story_translation(app, db, services):
-    story_translation = create_story_translation(db)
+    _, _, _, story_translation = create_story_translation(db)
     user_id = -92653
     assert User.query.get(user_id) == None
     with pytest.raises(Exception) as e:
